requirements_base.txt
Django==1.9.9
celery==3.1.23
amqp==2.0.3
anyjson==0.3.3
billiard==3.5.0.0
django-bootstrap3==7.0.1
decorator==4.0.10
feedparser==5.2.1
kombu==3.0.35
mock==2.0.0
oauthlib==1.1.2
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
ptyprocess==0.5.1
pytz==2016.6.1
requests==2.11.1
requests-oauthlib==0.6.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
twython==3.4.0
dropbox==6.4
gunicorn==19.6.0
psycopg2==2.6.2
django-allauth==0.27.0
django-nocaptcha-recaptcha==0.0.19
